js_compile: Don't shadow formal arguments with function name.
When a function's name is the same as one of its formal arguments, the
name should refer to the argument, but it was referring to the function.

Other than the obvious bug of referencing an incorrect value, it also
failed to compile in such case in strict mode (but only for lightweight
functions) with "duplicate formal parameter". On non lightweight (and
strict) it was "just" an incorrect value.